# docs/swagger.yaml
definitions:
  dtos.BuscarEnderecoResponse:
    properties:
      bairro:
        type: string
      cep:
        type: integer
      cidade:
        type: string
      estado:
        type: string
      logradouro:
        type: string
      numero:
        type: string
    type: object
  dtos.BuscarFotosResponse:
    properties:
      titulo:
        type: string
      url:
        type: string
    type: object
  dtos.BuscarListagemProfissionaisResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dtos.ProfissionalReduzido'
        type: array
      pagina:
        type: integer
      quantidade:
        type: integer
      total:
        type: integer
    type: object
  dtos.BuscarPodcastsResponse:
    properties:
      codigo:
        type: string
      titulo:
        type: string
    type: object
  dtos.BuscarProfissionalPorIdResponse:
    properties:
      celular:
        type: integer
      conselho:
        type: string
      email:
        type: string
      endereco_id:
        type: integer
      especialidades:
        items:
          type: string
        type: array
      facebook:
        type: string
      id:
        type: integer
      imagem_perfil_url:
        type: string
      instagram:
        type: string
      nome:
        type: string
      numero_identificacao:
        type: string
      site:
        type: string
      sobre:
        type: string
      telefone:
        type: integer
      tipo:
        type: string
    type: object
  dtos.BuscarUnidadesResponse:
    properties:
      ativo:
        type: boolean
      celular:
        type: integer
      descricao:
        type: string
      email:
        type: string
      endereco_id:
        type: integer
      facebook:
        type: string
      id:
        type: integer
      instagram:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      telefone:
        type: integer
      url_amigavel:
        type: string
      youtube:
        type: string
    type: object
  dtos.BuscarVideosResponse:
    properties:
      titulo:
        type: string
      url:
        type: string
      url_thumbnail:
        type: string
    type: object
  dtos.BuscarWhatsappsResponse:
    properties:
      numero:
        type: integer
      principal:
        type: boolean
    type: object
  dtos.ProfissionalReduzido:
    properties:
      email:
        type: string
      especialidades:
        items:
          type: string
        type: array
      facebook:
        type: string
      imagem_perfil_url:
        type: string
      instagram:
        type: string
      linkedin:
        type: string
      nome:
        type: string
      recomendado:
        type: boolean
      site:
        type: string
      tipo:
        type: string
      unidade_id:
        type: integer
      url_amigavel:
        type: string
      whatsapp:
        type: integer
      youtube:
        type: string
    type: object
info:
  contact:
    email: joaovitorpina@icloud.com
    name: João Vitor Goncalves Pina
  description: Essa é a API para o site do Informacão que Salva e o admin.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: IQS API
  version: "1.0"
paths:
  /admin/profissionais/{id}:
    get:
      description: Retorna todos os detalhes do profissional pelo id
      parameters:
        - description: Id do profissional
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: Profissional
          schema:
            $ref: '#/definitions/dtos.BuscarProfissionalPorIdResponse'
      summary: Busca todas as informacões do profissional
      tags:
        - Profissional
  /admin/profissionais/{id}/convenios:
    get:
      description: Buscar os convenios do profisional por id
      parameters:
        - description: Id do profissional
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: Convenios
          schema:
            items:
              type: string
            type: array
      summary: Busca os convenios do profissional
      tags:
        - Convenio
  /admin/profissionais/{id}/midias/fotos:
    get:
      description: Buscar as fotos do profisional por id
      parameters:
        - description: Id do profissional
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: Fotos
          schema:
            items:
              $ref: '#/definitions/dtos.BuscarFotosResponse'
            type: array
      summary: Busca as fotos do profissional
      tags:
        - Midias
  /admin/profissionais/{id}/midias/podcasts:
    get:
      description: Buscar os podcasts do profisional por id
      parameters:
        - description: Id do profissional
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: Podcasts
          schema:
            items:
              $ref: '#/definitions/dtos.BuscarPodcastsResponse'
            type: array
      summary: Busca os podcasts do profissional
      tags:
        - Midias
  /admin/profissionais/{id}/midias/videos:
    get:
      description: Buscar os videos do profisional por id
      parameters:
        - description: Id do profissional
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: Videos
          schema:
            items:
              $ref: '#/definitions/dtos.BuscarVideosResponse'
            type: array
      summary: Busca os videos do profissional
      tags:
        - Midias
  /admin/profissionais/{id}/tratamentos:
    get:
      description: Buscar os tratamentos do profisional por id
      parameters:
        - description: Id do profissional
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: Tratamentos
          schema:
            items:
              type: string
            type: array
      summary: Busca os tratamentos do profissional
      tags:
        - Tratamentos
  /admin/profissionais/{id}/whatsapps:
    get:
      description: Buscar os WhatsApp`s do profisional por id
      parameters:
        - description: Id do profissional
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: WhatsApp`s
          schema:
            items:
              $ref: '#/definitions/dtos.BuscarWhatsappsResponse'
            type: array
      summary: Busca os WhatsApp`s do profissional
      tags:
        - WhatsApp
  /enderecos/{id}:
    get:
      description: Retorna o endereco pelo id enviado
      parameters:
        - description: Id do Endereco
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: Endereco
          schema:
            $ref: '#/definitions/dtos.BuscarEnderecoResponse'
      summary: Busca um endereco por id
      tags:
        - Endereco
  /profissionais:
    get:
      description: Realiza a listagem com paginacão dos profissionais a partir dos
        dados enviados para busca
      parameters:
        - in: query
          name: especialidadeId
          type: integer
        - in: query
          name: limite
          type: integer
        - in: query
          name: nome
          type: string
        - in: query
          name: pagina
          type: integer
        - in: query
          name: tipoProfissionalId
          type: integer
        - in: query
          name: unidadeId
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: Listagem profissionais
          schema:
            $ref: '#/definitions/dtos.BuscarListagemProfissionaisResponse'
      summary: Listagem de profissionais pela busca
      tags:
        - Profissional
  /profissionais/{url_amigavel}:
    get:
      description: Retorna todos os detalhes do profissional pelo id
      parameters:
        - description: Url Amigavel do profissional
          in: path
          name: url_amigavel
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: Profissional
          schema:
            $ref: '#/definitions/dtos.BuscarProfissionalPorIdResponse'
      summary: Busca todas as informacões do profissional
      tags:
        - Profissional
  /profissionais/{url_amigavel}/convenios:
    get:
      description: Buscar os convenios do profisional por Url Amigavel
      parameters:
        - description: Url Amigavel do profissional
          in: path
          name: url_amigavel
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: Convenios
          schema:
            items:
              type: string
            type: array
      summary: Busca os convenios do profissional
      tags:
        - Convenio
  /profissionais/{url_amigavel}/midias/fotos:
    get:
      description: Buscar as fotos do profisional por Url Amigavel
      parameters:
        - description: Url Amigavel do profissional
          in: path
          name: url_amigavel
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: Fotos
          schema:
            items:
              $ref: '#/definitions/dtos.BuscarFotosResponse'
            type: array
      summary: Busca as fotos do profissional
      tags:
        - Midias
  /profissionais/{url_amigavel}/midias/podcasts:
    get:
      description: Buscar os podcasts do profisional por Url Amigavel
      parameters:
        - description: Url Amigavel do profissional
          in: path
          name: url_amigavel
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: Podcasts
          schema:
            items:
              $ref: '#/definitions/dtos.BuscarPodcastsResponse'
            type: array
      summary: Busca os podcasts do profissional
      tags:
        - Midias
  /profissionais/{url_amigavel}/midias/videos:
    get:
      description: Buscar os videos do profisional por Url Amigavel
      parameters:
        - description: Url Amigavel do profissional
          in: path
          name: url_amigavel
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: Videos
          schema:
            items:
              $ref: '#/definitions/dtos.BuscarVideosResponse'
            type: array
      summary: Busca os videos do profissional
      tags:
        - Midias
  /profissionais/{url_amigavel}/tratamentos:
    get:
      description: Buscar os tratamentos do profisional por Url Amigavel
      parameters:
        - description: Url Amigavel do profissional
          in: path
          name: url_amigavel
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: Tratamentos
          schema:
            items:
              type: string
            type: array
      summary: Busca os tratamentos do profissional
      tags:
        - Tratamentos
  /profissionais/{url_amigavel}/whatsapps:
    get:
      description: Buscar os WhatsApp`s do profisional por Url Amigavel
      parameters:
        - description: Url Amigavel do profissional
          in: path
          name: url_amigavel
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: WhatsApp`s
          schema:
            items:
              $ref: '#/definitions/dtos.BuscarWhatsappsResponse'
            type: array
      summary: Busca os WhatsApp`s do profissional
      tags:
        - WhatsApp
  /unidades:
    get:
      description: Retorna uma listagem com todas as unidades
      produces:
        - application/json
      responses:
        "200":
          description: Unidades
          schema:
            items:
              $ref: '#/definitions/dtos.BuscarUnidadesResponse'
            type: array
      summary: Lista todas as unidades
      tags:
        - Unidade
swagger: "2.0"